requirements.txt
# Core Dependencies
python-dateutil>=2.8.2
pytz>=2023.3
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
tensorflow>=2.13.0
keras>=2.13.0

# Financial Data
yfinance>=0.2.28
alpha-vantage>=2.3.1
pandas-datareader>=0.10.0

# Risk Metrics
arch>=6.2.0

# Explainability
shap>=0.42.0

# API Framework
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Dashboard
streamlit>=1.25.0
plotly>=5.16.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
joblib>=1.3.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# RAG Chatbot
langchain>=0.0.200
langchain-community>=0.4.0
langchain-openai>=1.0.0
langchain-groq>=0.2.0
langchain-core>=1.2.0
openai>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0
sentence-transformers>=3.0.0

